Fix saving user settings
Github api fails when creating zero length gist files, so if the content
is zero length, use "{}" instead which is fine for json based settings
files.

VSCode changed as part of the 1.4 merge to first create a zero length file
if it doesn't already exist before saving anything to it.
(see TextFileEditorModel.doSaveTextFileAs()).

Regular files are treated the same way. This may not be a problem - the
act of creating a new file (via explorer button) is a user action that
creates and commits a zero length file. Users can mentally tie this action
to the repo commit of a zero length file that results. The more important
issue is that currently the commit message says "Update <filename>".